[sparkle] Promote New* → canonical names, old → Legacy*
- NewButton → Button, old Button → LegacyButton
- NewSplitButton → SplitButton (FlexSplitButton), old → LegacyFlexSplitButton
- NewInput → Input, old Input → LegacyInput
- NewInputWithSave → InputWithSave, old → LegacyInputWithSave

index.ts exports updated. Stories renamed to match.
Intra-sparkle component imports left untouched.
