+++ sparkle/src/components/Button.tsx
@@ -5,370 +5,236 @@ import {
   LinkWrapper,
   type LinkWrapperProps,
 } from "@sparkle/components/LinkWrapper";
-import type { SpinnerProps } from "@sparkle/components/Spinner";
 import { Spinner } from "@sparkle/components/Spinner";
 import { Tooltip } from "@sparkle/components/Tooltip";
 import { ChevronDown } from "@sparkle/icons/v2-stroke";
 import { cn } from "@sparkle/lib/utils";
 import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
-import { useEffect, useMemo, useState } from "react";
 
-const PULSE_ANIMATION_DURATION = 1;
+// Redesigned button alongside the existing Button (unchanged). Sizes: 24/32/40px.
+// Dark-mode strategy per variant:
+//   - primary/outline: explicit dark: gradient overrides (primary goes lighter, outline darker)
+//   - highlight/warning: explicit dark: overrides to preserve hue (token ramp would flip to wrong shade)
+//   - ghost variants: semantic tokens auto-flip, no dark: needed
 
 export const BUTTON_VARIANTS = [
   "primary",
   "highlight",
-  "highlight-secondary",
-  "warning",
-  "warning-secondary",
   "outline",
+  "warning",
   "ghost",
   "ghost-secondary",
+  "highlight-ghost",
+  "warning-ghost",
 ] as const;
 
 export type ButtonVariantType = (typeof BUTTON_VARIANTS)[number];
 
-export const REGULAR_BUTTON_SIZES = [
-  "xmini",
-  "mini",
-  "xs",
-  "sm",
-  "md",
-] as const;
-export const ICON_ONLY_SIZES = ["icon-xs", "icon", "icon-sm"] as const;
-export const SMALL_BUTTON_SIZES = ["icon-xs", "icon", "xmini", "mini"] as const;
+export const BUTTON_SIZES = ["xs", "sm", "md"] as const;
+export type ButtonSizeType = (typeof BUTTON_SIZES)[number];
 
-export type RegularButtonSize = (typeof REGULAR_BUTTON_SIZES)[number];
-export type IconOnlySize = (typeof ICON_ONLY_SIZES)[number];
-export type ButtonSize = RegularButtonSize | IconOnlySize;
+// Inset white catch-light + hairline via border-dark + ambient drop via foreground.
+// All three auto-invert in dark mode via flipping tokens except the white highlight (intentional).
+const RAISED_SHADOW =
+  "shadow-[inset_0_0_1px_0_rgba(255,255,255,0.08),0_0_0.5px_0_var(--color-border-dark),0_1px_1.5px_0_color-mix(in_oklch,var(--color-foreground)_10%,transparent)]";
 
-function isSmallButtonSize(
-  size: ButtonSize | undefined
-): size is (typeof SMALL_BUTTON_SIZES)[number] {
-  return (
-    size !== undefined &&
-    SMALL_BUTTON_SIZES.includes(size as (typeof SMALL_BUTTON_SIZES)[number])
-  );
-}
+// Per-variant (not base) so ghost variants keep ::after free for consumers (e.g. Tabs).
+const OVERLAY = cn(
+  "after:pointer-events-none after:absolute after:inset-0 after:rounded-[inherit]",
+  "after:transition-colors disabled:after:hidden"
+);
 
-// Define button styling with cva
 const buttonVariants = cva(
   cn(
-    "inline-flex items-center justify-center whitespace-nowrap ring-offset-background transition-colors ring-inset select-none",
-    "focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-0"
+    "relative isolate inline-flex shrink-0 select-none items-center justify-center whitespace-nowrap",
+    // `transform` stays in the transition list for the `press` scale.
+    "transition-[color,background-color,border-color,transform] duration-100 ease-out",
+    "motion-reduce:transition-none",
+    "focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-0",
+    // data-disabled is set only when truly disabled (not loading), so loading buttons
+    // keep their full-color active look while :disabled guards block interaction for both.
+    "data-[disabled]:shadow-none",
+    "disabled:cursor-not-allowed disabled:focus-visible:ring-0"
   ),
   {
     variants: {
       variant: {
         primary: cn(
-          "border border-transparent",
-          "bg-primary-800",
+          OVERLAY,
+          "bg-linear-to-b from-primary-700 to-primary-800",
+          "dark:from-stone-50 dark:to-stone-150",
           "text-primary-50",
-          "hover:bg-primary-light",
-          "active:bg-primary-dark",
-          "disabled:bg-primary-muted disabled:text-highlight-on/60"
+          RAISED_SHADOW,
+          "data-[disabled]:from-primary-300 data-[disabled]:to-primary-400",
+          "dark:data-[disabled]:opacity-50",
+          // Dark: light button surface → darken on hover instead of lighten.
+          // Dark: light button → drop shadow uses black (not white foreground token) to avoid glow.
+          "dark:shadow-[inset_0_0_1px_0_rgba(255,255,255,0.5),0_0_0.5px_0_var(--color-border-dark),0_1px_1.5px_0_rgba(0,0,0,0.1)]",
+          "dark:hover:after:bg-black/[0.04] dark:active:after:bg-black/[0.04]"
         ),
         highlight: cn(
-          "border border-transparent",
-          "bg-highlight",
-          "text-highlight-on",
-          "hover:bg-highlight-light",
-          "active:bg-highlight-dark",
-          "disabled:bg-highlight-muted disabled:text-highlight-on/60"
-        ),
-        "highlight-secondary": cn(
-          "border",
-          "border-border",
-          "text-highlight-500",
-          "bg-background",
-          "hover:text-highlight-500",
-          "hover:bg-highlight-50",
-          "hover:border-primary-150",
-          "active:bg-primary-300",
-          "disabled:text-primary-muted",
-          "disabled:border-primary-100",
-          "disabled:hover:bg-background",
-          "disabled:hover:border-primary-100",
-          "disabled:hover:text-primary-muted"
+          OVERLAY,
+          "bg-linear-to-b from-highlight-400 to-highlight-500",
+          "dark:from-blue-500 dark:to-blue-600",
+          "text-white",
+          RAISED_SHADOW,
+          "dark:shadow-[inset_0_0_1px_0_rgba(255,255,255,0.08),0_0_0.5px_0_var(--color-border-dark),0_1px_1.5px_0_rgba(0,0,0,0.1)]",
+          "data-[disabled]:from-highlight-200 data-[disabled]:to-highlight-300",
+          // Dark disabled: lock gradient (prevent ramp flip) then fade with opacity.
+          "dark:data-[disabled]:from-highlight-600 dark:data-[disabled]:to-highlight-500",
+          "dark:data-[disabled]:opacity-50"
         ),
         warning: cn(
-          "border border-transparent",
-          "bg-warning",
-          "text-warning-on",
-          "hover:bg-warning-light",
-          "active:bg-warning-dark",
-          "disabled:bg-warning-muted disabled:text-highlight-on/60"
-        ),
-        "warning-secondary": cn(
-          "border",
-          "border-border",
-          "text-warning-500",
-          "bg-background",
-          "hover:text-warning-500",
-          "hover:bg-warning-50",
-          "hover:border-primary-150",
-          "active:bg-primary-300",
-          "disabled:text-primary-muted",
-          "disabled:border-primary-100",
-          "disabled:hover:bg-background",
-          "disabled:hover:border-primary-100",
-          "disabled:hover:text-primary-muted"
+          OVERLAY,
+          "bg-linear-to-b from-warning-400 to-warning-500",
+          "dark:from-red-500 dark:to-red-600",
+          "text-white",
+          RAISED_SHADOW,
+          "dark:shadow-[inset_0_0_1px_0_rgba(255,255,255,0.08),0_0_0.5px_0_var(--color-border-dark),0_1px_1.5px_0_rgba(0,0,0,0.1)]",
+          "data-[disabled]:from-warning-200 data-[disabled]:to-warning-300",
+          // Dark disabled: lock gradient (prevent ramp flip) then fade with opacity.
+          "dark:data-[disabled]:from-warning-600 dark:data-[disabled]:to-warning-500",
+          "dark:data-[disabled]:opacity-50"
         ),
         outline: cn(
-          "border",
-          "border-border",
-          "text-primary",
-          "bg-background",
-          "hover:text-primary",
-          "hover:bg-primary-100",
-          "hover:border-primary-150",
-          "active:bg-primary-300",
-          "disabled:text-primary-muted",
-          "disabled:border-primary-100",
-          "disabled:hover:bg-background",
-          "disabled:hover:border-primary-100",
-          "disabled:hover:text-primary-muted"
+          OVERLAY,
+          // overflow-hidden + after:rounded-none clips the hover overlay flush to the border-radius.
+          "overflow-hidden after:rounded-none",
+          "border border-border-dark dark:border-stone-775",
+          "bg-linear-to-b from-primary-50 to-primary-100",
+          "dark:from-stone-800 dark:to-stone-800",
+          "text-muted-foreground dark:text-primary-900",
+          "shadow-[inset_0_0_1px_0_rgba(255,255,255,0.08),0_0_0.5px_0_var(--color-border-dark),0_0.5px_1px_0_rgba(0,0,0,0.06)]",
+          "hover:after:bg-black/[0.02] active:after:bg-black/[0.02]",
+          "data-[disabled]:text-faint",
+          "dark:data-[disabled]:opacity-50"
         ),
         ghost: cn(
-          "border",
-          "border-transparent",
           "text-foreground",
-          "hover:bg-hover",
-          "hover:text-primary-900",
-          "hover:border-transparent",
-          "active:bg-primary-300",
-          "disabled:text-faint",
-          "disabled:hover:bg-transparent",
-          "disabled:hover:border-transparent",
-          "disabled:hover:text-faint"
+          "hover:bg-black/[0.02] active:bg-black/[0.02]",
+          "dark:hover:bg-white/[0.08] dark:active:bg-white/[0.08]",
+          "data-[disabled]:text-faint",
+          "disabled:hover:bg-transparent dark:disabled:hover:bg-transparent"
         ),
         "ghost-secondary": cn(
-          "border",
-          "border-transparent",
           "text-muted-foreground",
-          "hover:bg-hover",
-          "hover:text-primary-900",
-          "hover:border-transparent",
-          "active:bg-primary-300",
-          "disabled:text-faint",
-          "disabled:hover:bg-transparent",
-          "disabled:hover:border-transparent",
-          "disabled:hover:text-faint"
+          "hover:bg-black/[0.02] active:bg-black/[0.02]",
+          "dark:hover:bg-white/[0.08] dark:active:bg-white/[0.08]",
+          "data-[disabled]:text-faint",
+          "disabled:hover:bg-transparent dark:disabled:hover:bg-transparent"
+        ),
+        "highlight-ghost": cn(
+          "text-highlight-500",
+          "hover:bg-highlight-50 active:bg-highlight-50",
+          "data-[disabled]:text-highlight-muted",
+          "disabled:hover:bg-transparent"
+        ),
+        "warning-ghost": cn(
+          "text-warning-500",
+          "hover:bg-warning-50 active:bg-warning-50",
+          "data-[disabled]:text-warning-muted",
+          "disabled:hover:bg-transparent"
         ),
       },
       size: {
-        "icon-xs": "h-6 w-6 gap-1 shrink-0",
-        icon: "h-7 w-7 gap-1.5 shrink-0",
-        "icon-sm": "h-9 w-9 gap-2 shrink-0",
-        xmini: "h-6 px-1.5 gap-1 shrink-0",
-        mini: "h-7 px-2 gap-1.5 shrink-0",
-        xs: "h-7 px-2.5 gap-1.5 shrink-0",
-        sm: "h-9 px-3 gap-2 shrink-0",
-        md: "h-12 px-4 py-2 gap-2.5 shrink-0",
+        xs: "h-6 gap-1.5 px-2 text-sm font-medium leading-4 tracking-[-0.28px] rounded-[9px]",
+        sm: "h-8 gap-1.5 px-3 text-sm font-medium tracking-[-0.28px] rounded-xl",
+        md: "h-10 gap-1.5 px-4 text-base font-medium tracking-[-0.32px] rounded-[15px]",
       },
-      rounded: {
-        "icon-xs": "rounded-lg",
-        icon: "rounded-lg",
-        "icon-sm": "rounded-xl",
-        xmini: "rounded-lg",
-        mini: "rounded-lg",
-        xs: "rounded-lg",
-        sm: "rounded-xl",
-        md: "rounded-2xl",
-        full: "rounded-full",
+      isIconOnly: {
+        true: "",
+        false: "",
+      },
+      press: {
+        true: "active:scale-[0.985] motion-reduce:active:scale-100",
+        false: "",
       },
     },
+    compoundVariants: [
+      { size: "xs", isIconOnly: true, className: "w-6 px-0" },
+      { size: "sm", isIconOnly: true, className: "w-8 px-0" },
+      { size: "md", isIconOnly: true, className: "w-10 px-0" },
+      // Solid raised variants: larger buttons get a stronger white overlay (Figma spec).
+      // dark:hover on primary is overridden per-variant to darken instead of lighten.
+      {
+        variant: ["primary", "highlight", "warning"],
+        size: ["xs", "sm"],
+        className: "hover:after:bg-white/10 active:after:bg-white/10",
+      },
+      {
+        variant: ["primary", "highlight", "warning"],
+        size: "md",
+        className: "hover:after:bg-white/20 active:after:bg-white/20",
+      },
+      // Outline dark: white overlay (dark surface needs lighten, not darken).
+      {
+        variant: "outline",
+        size: ["xs", "sm", "md"],
+        className:
+          "dark:hover:after:bg-white/[0.02] dark:active:after:bg-white/[0.02]",
+      },
+    ],
     defaultVariants: {
       variant: "primary",
       size: "sm",
-      rounded: "sm",
+      isIconOnly: false,
+      press: true,
     },
   }
 );
 
-const labelVariants = cva("", {
-  variants: {
-    size: {
-      "icon-xs": "hidden",
-      icon: "hidden",
-      "icon-sm": "hidden",
-      xmini: "",
-      mini: "",
-      xs: "",
-      sm: "",
-      md: "",
-    },
-    hasLighterFont: {
-      true: "",
-      false: "",
-    },
-  },
-  compoundVariants: [
-    { size: "xmini", hasLighterFont: false, className: "label-xs" },
-    { size: "mini", hasLighterFont: false, className: "label-xs" },
-    { size: "xs", hasLighterFont: false, className: "label-xs" },
-    { size: "sm", hasLighterFont: false, className: "label-sm" },
-    { size: "md", hasLighterFont: false, className: "label-base" },
-    {
-      size: "xmini",
-      hasLighterFont: true,
-      className: "text-xs font-normal",
-    },
-    {
-      size: "mini",
-      hasLighterFont: true,
-      className: "text-xs font-normal",
-    },
-    { size: "xs", hasLighterFont: true, className: "text-xs font-normal" },
-    { size: "sm", hasLighterFont: true, className: "text-sm font-normal" },
-    {
-      size: "md",
-      hasLighterFont: true,
-      className: "text-base font-normal",
-    },
-  ],
-  defaultVariants: {
-    size: "sm",
-  },
-});
+// Labels and icons carry a subtle shadow on raised (non-ghost) variants.
+const RAISED_VARIANTS: ButtonVariantType[] = [
+  "primary",
+  "highlight",
+  "warning",
+  "outline",
+];
+const TEXT_SHADOW = "[text-shadow:0_1px_1.5px_rgba(0,0,0,0.08)]";
+const ICON_SHADOW = "drop-shadow-[0px_1px_0.75px_rgba(0,0,0,0.08)]";
 
-type SpinnerVariant = NonNullable<SpinnerProps["variant"]>;
-
-const spinnerVariantsMap: Record<ButtonVariantType, SpinnerVariant> = {
-  primary: "revert",
-  highlight: "light",
-  "highlight-secondary": "mono",
-  warning: "light",
-  "warning-secondary": "mono",
-  outline: "mono",
-  ghost: "mono",
-  "ghost-secondary": "mono",
+const ICON_SIZE_MAP: Record<ButtonSizeType, "xs" | "sm"> = {
+  xs: "xs",
+  sm: "xs",
+  md: "sm",
 };
 
-const chevronVariantMap = {
-  primary: "text-faint",
+const COUNTER_SIZE_MAP: Record<ButtonSizeType, "xs" | "sm"> = {
+  xs: "xs",
+  sm: "xs",
+  md: "sm",
+};
+
+const chevronVariantMap: Record<ButtonVariantType, string> = {
+  // primary swaps to a light button in dark mode, so its chevron tracks the
+  // flipping text token rather than a fixed white.
+  primary: "text-primary-50/60",
+  highlight: "text-white/60",
+  warning: "text-white/60",
   outline: "text-faint",
   ghost: "text-faint",
   "ghost-secondary": "text-faint",
-  highlight: "text-white/60",
-  "highlight-secondary": "text-highlight-500",
-  warning: "text-white/60",
-  "warning-secondary": "text-warning-500",
-} as const;
-
-export interface MetaButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
-  isRounded?: boolean;
-}
-
-const MetaButton = React.forwardRef<HTMLButtonElement, MetaButtonProps>(
-  (
-    {
-      className,
-      asChild = false,
-      variant,
-      size,
-      isRounded,
-      children,
-      ...props
-    },
-    ref
-  ) => {
-    const Comp = asChild ? Slot : "button";
-
-    // Determine rounded variant based on isRounded prop
-    const rounded = isRounded ? "full" : size;
-
-    return (
-      <Comp
-        className={cn(buttonVariants({ variant, size, rounded, className }))}
-        ref={ref}
-        {...props}
-      >
-        {children}
-      </Comp>
-    );
-  }
-);
-MetaButton.displayName = "MetaButton";
-
-type IconSizeType = "xs" | "sm" | "md";
-type CounterSizeType = "xs" | "sm" | "md";
-
-export const ICON_SIZE_MAP: Record<ButtonSize, IconSizeType> = {
-  "icon-xs": "xs",
-  icon: "sm",
-  "icon-sm": "sm",
-  xmini: "xs",
-  mini: "sm",
-  xs: "xs",
-  sm: "sm",
-  md: "md",
+  "highlight-ghost": "text-highlight-400",
+  "warning-ghost": "text-warning-400",
 };
 
-const COUNTER_SIZE_MAP: Record<ButtonSize, CounterSizeType> = {
-  "icon-xs": "xs",
-  icon: "xs",
-  "icon-sm": "sm",
-  xmini: "xs",
-  mini: "xs",
-  xs: "xs",
-  sm: "sm",
-  md: "md",
+// Loading spinner color, matched to each variant's text color (including the
+// dark-mode swap: `revert` = light spinner on light theme, `mono` = the
+// inverse). Keeps the spinner the same color as the label.
+const spinnerVariantMap: Record<
+  ButtonVariantType,
+  React.ComponentProps<typeof Spinner>["variant"]
+> = {
+  primary: "revert",
+  highlight: "light",
+  warning: "light",
+  outline: "mono",
+  ghost: "mono",
+  "ghost-secondary": "mono",
+  "highlight-ghost": "blue500",
+  "warning-ghost": "red500",
 };
-
-const loadingContainerVariants = cva("-mx-0.5", {
-  variants: {
-    size: {
-      "icon-xs": "w-5 px-0.5",
-      icon: "w-5 px-0.5",
-      "icon-sm": "",
-      xmini: "w-5 px-0.5",
-      mini: "w-5 px-0.5",
-      xs: "w-5 px-0.5",
-      sm: "",
-      md: "",
-    },
-  },
-  defaultVariants: {
-    size: "sm",
-  },
-});
-
-const selectButtonSizeVariants = cva("", {
-  variants: {
-    size: {
-      "icon-xs": "w-auto px-1.5",
-      xmini: "w-auto px-1.5",
-      mini: "w-auto px-2",
-      icon: "w-auto px-2",
-      "icon-sm": "",
-      xs: "",
-      sm: "",
-      md: "",
-    },
-  },
-  defaultVariants: {
-    size: "sm",
-  },
-});
-
-type CommonButtonProps = Omit<MetaButtonProps, "children"> &
-  Omit<LinkWrapperProps, "children"> & {
-    isSelect?: boolean;
-    isLoading?: boolean;
-    isPulsing?: boolean;
-    briefPulse?: boolean;
-    tooltip?: string;
-    tooltipShortcut?: string;
-    isCounter?: boolean;
-    counterValue?: string;
-    isRounded?: boolean;
-    hasLighterFont?: boolean;
-  };
 
 export type ButtonIconType = React.ComponentType | React.ReactElement;
 
@@ -376,38 +242,39 @@ function isReactElement(visual: ButtonIconType): visual is React.ReactElement {
   return React.isValidElement(visual);
 }
 
-export type IconOnlyButtonProps = CommonButtonProps & {
-  size: IconOnlySize;
-  icon: ButtonIconType;
-  label?: never;
-};
-
-export type RegularButtonProps = CommonButtonProps & {
-  size?: RegularButtonSize;
-  icon?: ButtonIconType;
+export interface ButtonProps
+  extends Omit<React.ButtonHTMLAttributes<HTMLButtonElement>, "size">,
+    Omit<LinkWrapperProps, "children" | "className">,
+    Pick<VariantProps<typeof buttonVariants>, "variant"> {
+  size?: ButtonSizeType;
   label?: string;
-};
+  icon?: ButtonIconType;
+  iconRight?: ButtonIconType;
+  isSelect?: boolean;
+  isLoading?: boolean;
+  isCounter?: boolean;
+  counterValue?: string;
+  isPulsing?: boolean;
+  tooltip?: string;
+  tooltipShortcut?: string;
+}
 
-export type ButtonProps = IconOnlyButtonProps | RegularButtonProps;
-
-const LegacyButton = React.forwardRef<HTMLButtonElement, ButtonProps>(
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   (
     {
       label,
       icon,
+      iconRight,
       className,
-      isLoading = false,
       variant = "primary",
-      tooltip,
-      tooltipShortcut,
+      size = "sm",
       isSelect = false,
-      isPulsing = false,
-      briefPulse = false,
+      isLoading = false,
       isCounter = false,
       counterValue,
-      size = "sm",
-      isRounded = false,
-      hasLighterFont = false,
+      isPulsing = false,
+      tooltip,
+      tooltipShortcut,
       href,
       target,
       rel,
@@ -419,80 +286,55 @@ const LegacyButton = React.forwardRef<HTMLButtonElement, ButtonProps>(
     ref
   ) => {
     const iconSize = ICON_SIZE_MAP[size];
-    const counterSize = COUNTER_SIZE_MAP[size];
-
-    const [isPulsingBriefly, setIsPulsingBriefly] = useState(false);
-
-    useEffect(() => {
-      if (!briefPulse) {
-        return;
-      }
-      const startPulse = () => {
-        setIsPulsingBriefly(true);
-        setTimeout(
-          () => setIsPulsingBriefly(false),
-          PULSE_ANIMATION_DURATION * 3000
-        );
-      };
-      startPulse();
-    }, [briefPulse]);
+    const showCounter = isCounter && counterValue != null;
+    const isIconOnly = !label && !showCounter && !isSelect && !!icon;
+    // Menu triggers (dropdown/popover/select) skip the press scale, else the
+    // opening menu jumps with the anchor. Radix sets aria-haspopup on the
+    // trigger; isSelect is our chevron affordance. Tooltips don't, so they keep it.
+    const isMenuTrigger = isSelect || props["aria-haspopup"] != null;
+    const hasTextShadow = variant != null && RAISED_VARIANTS.includes(variant);
+    const iconShadow = hasTextShadow ? ICON_SHADOW : "";
 
     const renderIcon = (visual: ButtonIconType, extraClass = "") => {
       if (isReactElement(visual)) {
-        return <span className={cn(extraClass, "shrink-0")}>{visual}</span>;
+        return <span className={cn("shrink-0", extraClass)}>{visual}</span>;
       }
-
-      return (
-        <Icon visual={visual} size={iconSize} className={cn(extraClass)} />
-      );
+      return <Icon visual={visual} size={iconSize} className={extraClass} />;
     };
-    const renderChevron = (visual: React.ComponentType, extraClass = "") => (
-      <Icon
-        visual={visual}
-        size={iconSize}
-        className={cn(variant ? chevronVariantMap[variant] : "", extraClass)}
-      />
-    );
-
-    const showCounter = isCounter && counterValue != null;
-    const showContainer = label || showCounter;
 
     const content = (
       <>
         {isLoading ? (
-          <div className={loadingContainerVariants({ size })}>
-            <Spinner
-              size={isSmallButtonSize(size) ? "xs" : iconSize}
-              variant={(variant && spinnerVariantsMap[variant]) || "gray400"}
-            />
-          </div>
+          <Spinner
+            size="xs"
+            variant={spinnerVariantMap[variant ?? "primary"]}
+          />
         ) : (
-          icon && renderIcon(icon, "-mx-0.5")
+          icon && renderIcon(icon, iconShadow)
         )}
-
-        {showContainer && (
-          <div
-            className={cn(
-              "flex items-center gap-2",
-              labelVariants({ size, hasLighterFont })
-            )}
-          >
-            {label}
-            {showCounter && (
-              <Counter
-                value={Number(counterValue)}
-                variant={variant || "primary"}
-                size={counterSize}
-                isInButton={true}
-              />
-            )}
-          </div>
+        {label && (
+          <span className={cn(hasTextShadow && TEXT_SHADOW)}>{label}</span>
         )}
-        {isSelect && renderChevron(ChevronDown, isLoading ? "" : "-mr-1")}
+        {showCounter && (
+          <Counter
+            value={Number(counterValue)}
+            variant="primary"
+            size={COUNTER_SIZE_MAP[size]}
+            isInButton={true}
+          />
+        )}
+        {!isLoading && iconRight && renderIcon(iconRight, iconShadow)}
+        {isSelect && (
+          <Icon
+            visual={ChevronDown}
+            size={iconSize}
+            className={variant ? chevronVariantMap[variant] : ""}
+          />
+        )}
       </>
     );
 
-    const pointerEventProps = useMemo(() => {
+    const pointerEventProps = React.useMemo(() => {
       if (isLoading || props.disabled) {
         return {
           onPointerDown: (e: React.PointerEvent<HTMLButtonElement>) => {
@@ -504,32 +346,39 @@ const LegacyButton = React.forwardRef<HTMLButtonElement, ButtonProps>(
       return {};
     }, [isLoading, props.disabled]);
 
-    // We cannot skip a button tag when it's disabled. We need
-    // to apply disabled class manually (currently it has :disabled pseudo-class, which won't work if it's not a button)
-    // and disable pointer events.
-    const shouldUseSlot = !!href && !props.disabled;
-
-    const innerContent = shouldUseSlot ? <span>{content}</span> : content;
+    // Disabled/loading buttons must not navigate: render a real <button> and
+    // skip the LinkWrapper below.
+    const isInteractive = !props.disabled && !isLoading;
+    const shouldUseSlot = !!href && isInteractive;
+    const Comp = shouldUseSlot ? Slot : "button";
 
     const innerButton = (
-      <MetaButton
+      <Comp
         ref={ref}
-        size={size}
-        variant={variant}
-        isRounded={isRounded}
-        disabled={isLoading || props.disabled}
         className={cn(
-          (isPulsing || isPulsingBriefly) && "animate-ring-pulse",
-          isSelect && selectButtonSizeVariants({ size }),
+          buttonVariants({
+            variant,
+            size,
+            isIconOnly,
+            press: !isMenuTrigger && !isLoading,
+          }),
+          isPulsing && "animate-ring-pulse-soft",
           className
         )}
         aria-label={ariaLabel || tooltip || label}
-        asChild={shouldUseSlot}
         {...props}
+        // Loading blocks interaction (disabled attr) but keeps the active look:
+        // the muted `data-[disabled]:` styles are gated on this attribute, which
+        // is set only when truly disabled — so a loading button stays full-color
+        // and its spinner matches the label. Set after {...props} so a
+        // consumer-passed `disabled` can't clobber the computed value.
+        disabled={isLoading || props.disabled}
+        data-disabled={props.disabled && !isLoading ? "" : undefined}
+        aria-busy={isLoading || undefined}
         {...pointerEventProps}
       >
-        {innerContent}
-      </MetaButton>
+        {shouldUseSlot ? <span>{content}</span> : content}
+      </Comp>
     );
 
     const wrappedContent = tooltip ? (
@@ -543,7 +392,7 @@ const LegacyButton = React.forwardRef<HTMLButtonElement, ButtonProps>(
       innerButton
     );
 
-    return href ? (
+    return href && isInteractive ? (
       <LinkWrapper
         href={href}
         target={target}
@@ -559,17 +408,6 @@ const LegacyButton = React.forwardRef<HTMLButtonElement, ButtonProps>(
   }
 );
 
-LegacyButton.displayName = "LegacyButton";
+Button.displayName = "Button";
 
-export type LegacyButtonProps = ButtonProps;
-export type LegacyButtonVariantType = ButtonVariantType;
-export type LegacyRegularButtonSize = RegularButtonSize;
-export type LegacyRegularButtonProps = RegularButtonProps;
-export type LegacyIconOnlyButtonProps = IconOnlyButtonProps;
-export {
-  buttonVariants,
-  ICON_SIZE_MAP as LEGACY_ICON_SIZE_MAP,
-  LegacyButton,
-  LegacyButton as Button,
-  MetaButton,
-};
+export { Button, buttonVariants };
