+++ sparkle/src/stories/Input.stories.tsx
@@ -1,7 +1,9 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React from "react";
 
-import { Input } from "../index_with_tw_base";
+import { INPUT_SIZES } from "@sparkle/components/Input";
+
+import { Image03, Input, SearchMd } from "../index_with_tw_base";
 
 const MESSAGE_STATUSES = ["info", "default", "error"] as const;
 
@@ -27,6 +29,11 @@ const meta = {
     },
   },
   argTypes: {
+    size: {
+      options: INPUT_SIZES,
+      control: { type: "select" },
+      description: "The size of the input (sm / md / lg)",
+    },
     placeholder: {
       control: "text",
       description: "Placeholder text for the input",
@@ -69,6 +76,7 @@ type Story = StoryObj<typeof meta>;
 
 export const Playground: Story = {
   args: {
+    size: "md",
     placeholder: "Enter text...",
     value: "",
     label: "Input Label",
@@ -78,6 +86,41 @@ export const Playground: Story = {
     isError: false,
     type: "text",
   },
+};
+
+export const Sizes: Story = {
+  render: () => (
+    <div className="flex w-72 flex-col gap-4">
+      {INPUT_SIZES.map((size) => (
+        <Input key={size} size={size} placeholder={`Size ${size}`} />
+      ))}
+    </div>
+  ),
+};
+
+export const WithIcons: Story = {
+  render: () => (
+    <div className="flex w-72 flex-col gap-4">
+      <Input placeholder="Search…" icon={SearchMd} />
+      <Input placeholder="With trailing icon" iconRight={Image03} />
+      <Input placeholder="Both" icon={SearchMd} iconRight={Image03} />
+    </div>
+  ),
+};
+
+export const WithPrefixSuffix: Story = {
+  render: () => (
+    <div className="flex w-72 flex-col gap-4">
+      <Input
+        placeholder="0.00"
+        prefix={<span className="text-faint">$</span>}
+      />
+      <Input
+        placeholder="12,890"
+        suffix={<span className="text-faint">cr</span>}
+      />
+    </div>
+  ),
 };
 
 export const WithError: Story = {
