+++ sparkle/src/stories/SplitButton.stories.tsx
@@ -1,8 +1,9 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React from "react";
 
-import { Button, FlexSplitButton } from "../index_with_tw_base";
 import { ArrowUp, ChevronDown } from "@sparkle/icons/v2-stroke";
+
+import { Button, FlexSplitButton } from "../index_with_tw_base";
 
 const meta: Meta<React.ComponentProps<typeof FlexSplitButton>> = {
   title: "Actions/SplitButton",
@@ -10,15 +11,14 @@ const meta: Meta<React.ComponentProps<typeof FlexSplitButton>> = {
   parameters: {
     docs: {
       description: {
-        component: `A primary action paired with an attached secondary affordance. **FlexSplitButton** renders a labelled button (with \`label\`, \`icon\`, \`variant\`, and an \`isLoading\` state) joined to a \`splitAction\` — typically a chevron **Button** that opens a menu of related options.
+        component: `A primary action paired with an attached secondary affordance. **FlexSplitButton** renders a labelled **Button** (with \`label\`, \`icon\`, \`variant\`, and an \`isLoading\` state) joined to a \`splitAction\` — typically a chevron **Button** that opens a menu of related options. An opaque divider separates the two so the main button's hover/active overlay doesn't bleed through it.
 
 **When to use**
 - When one action is the obvious default but a few related variants should be one click away (e.g. "Send" + send options).
 
 **Guidelines**
 - Match the \`variant\` of the main button and the \`splitAction\` button so they read as one control.
-- Keep the main \`label\` as the most common action; relegate alternatives to the split menu.
-- For a set of equally-weighted related actions, use **ButtonGroup** instead.`,
+- Use an icon-only \`xs\` Button as the \`splitAction\`.`,
       },
     },
   },
@@ -27,156 +27,50 @@ const meta: Meta<React.ComponentProps<typeof FlexSplitButton>> = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const FlexSplitButtonVariants: Story = {
+const VARIANTS = [
+  "primary",
+  "highlight",
+  "outline",
+  "warning",
+  "ghost",
+  "ghost-secondary",
+  "highlight-ghost",
+  "warning-ghost",
+] as const;
+
+export const Variants: Story = {
   render: () => (
-    <div className="flex gap-3">
-      <FlexSplitButton
-        label="Send"
-        variant="highlight"
-        icon={ArrowUp}
-        splitAction={
-          <Button size="icon" variant="highlight" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="primary"
-        icon={ArrowUp}
-        splitAction={
-          <Button size="icon" variant="primary" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="outline"
-        icon={ArrowUp}
-        splitAction={
-          <Button size="icon" variant="outline" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="highlight-secondary"
-        icon={ArrowUp}
-        splitAction={
-          <Button
-            size="icon"
-            variant="highlight-secondary"
-            icon={ChevronDown}
-          />
-        }
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="warning"
-        icon={ArrowUp}
-        splitAction={
-          <Button size="icon" variant="warning" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="warning-secondary"
-        icon={ArrowUp}
-        splitAction={
-          <Button size="icon" variant="warning-secondary" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="ghost"
-        icon={ArrowUp}
-        splitAction={<Button size="icon" variant="ghost" icon={ChevronDown} />}
-      />
-      <FlexSplitButton
-        label="Send"
-        variant="ghost-secondary"
-        icon={ArrowUp}
-        splitAction={
-          <Button size="icon" variant="ghost-secondary" icon={ChevronDown} />
-        }
-      />
+    <div className="flex flex-wrap gap-3">
+      {VARIANTS.map((variant) => (
+        <FlexSplitButton
+          key={variant}
+          label="Send"
+          variant={variant}
+          icon={ArrowUp}
+          splitAction={
+            <Button size="xs" variant={variant} icon={ChevronDown} />
+          }
+        />
+      ))}
     </div>
   ),
 };
 
-export const FlexSplitButtonLoading: Story = {
+export const Loading: Story = {
   render: () => (
-    <div className="flex gap-3">
-      <FlexSplitButton
-        label="Sending"
-        variant="highlight"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button size="icon" variant="highlight" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="primary"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button size="icon" variant="primary" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="outline"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button size="icon" variant="outline" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="highlight-secondary"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button
-            size="icon"
-            variant="highlight-secondary"
-            icon={ChevronDown}
-          />
-        }
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="warning"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button size="icon" variant="warning" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="warning-secondary"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button size="icon" variant="warning-secondary" icon={ChevronDown} />
-        }
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="ghost"
-        icon={ArrowUp}
-        isLoading
-        splitAction={<Button size="icon" variant="ghost" icon={ChevronDown} />}
-      />
-      <FlexSplitButton
-        label="Sending"
-        variant="ghost-secondary"
-        icon={ArrowUp}
-        isLoading
-        splitAction={
-          <Button size="icon" variant="ghost-secondary" icon={ChevronDown} />
-        }
-      />
+    <div className="flex flex-wrap gap-3">
+      {(["highlight", "primary", "outline"] as const).map((variant) => (
+        <FlexSplitButton
+          key={variant}
+          label="Sending"
+          variant={variant}
+          icon={ArrowUp}
+          isLoading
+          splitAction={
+            <Button size="xs" variant={variant} icon={ChevronDown} />
+          }
+        />
+      ))}
     </div>
   ),
 };
